education/member/route

+++ app/Http/Controllers/Api/EducationController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 use Laravel\Lumen\Routing\Controller as BaseController;
 use Illuminate\Http\Request;
 use App\Models\Education;
+use App\Models\Groups;
 
 class EducationController extends BaseController
 {
@@ -19,23 +20,8 @@ class EducationController extends BaseController
         return response()->json($results);
     }
 
-    public function listEducation_ById($facultyId){
-        $results = Education::where('educationId',$facultyId)->get();
-
-        return response()->json($results);
-    }
-
-    public function Search_Education(Request $request){
-        $results = Education::where('educationName','like','%'.$request->educationName.'%')
-                            ->get();
-
-        return response()->json($results);
-    }
-
-    public function Search_Education_Active(Request $request){
-        $results = Education::where('educationName','like','%'.$request->educationName.'%')
-                            ->where('status','1')
-                            ->get();
+    public function listEducation_ById($educationId){
+        $results = Education::where('educationId',$educationId)->get();
 
         return response()->json($results);
     }
@@ -60,19 +46,34 @@ class EducationController extends BaseController
     }
 
     public function changstatusEducation(Request $request){
-        Education::where('educationId',$request->educationId)
-            ->update([
-                'status' => $request->status
-            ]);
-
-        $resp = array('status'=>1, 'message'=>'Change Status Success');
-        return response()->json($request);
+        $group = Education::join('groups','groups.educationLevel','=','education.educationId')
+                            ->where('education.educationId',$request->educationId)->get();
+        $subject = Education::join('subjects','subjects.educationLevel','=','education.educationId')
+                            ->where('education.educationId',$request->educationId)->get();
+        $resp = array('status'=>1, 'message'=>'');
+            if($subject == '[]' && $subject == '[]'){
+                Education::where('educationId',$request->educationId)
+                            ->update(['status' => $request->status]);
+                $resp = array('status'=>1, 'message'=>'เปลี่ยนสถานะระดับการศึกษาสำเร็จ');
+            }else{
+                $resp = array('status'=>0, 'message'=>'เปลี่ยนสถานะระดับการศึกษาไม่สำเร็จ');
+            }
+        return response()->json($resp);
     }
 
     public function deleteEducation(Request $request){
-        $deleteFaculty = Education::where('educationId',$request->educationId)->delete();
-
-        return response()->json($request);
+        $group = Education::join('groups','groups.educationLevel','=','education.educationId')
+                            ->where('education.educationId',$request->educationId)->get();
+        $subject = Education::join('subjects','subjects.educationLevel','=','education.educationId')
+                            ->where('education.educationId',$request->educationId)->get();
+        $resp = array('status'=>1, 'message'=>'');
+            if($subject == '[]' && $subject == '[]'){
+                $deleteFaculty = Education::where('educationId',$request->educationId)->delete();
+                $resp = array('status'=>1, 'message'=>'ลบระดับการศึกษาสำเร็จ');
+            }else{
+                $resp = array('status'=>0, 'message'=>'ลบระดับการศึกษาไม่สำเร็จ');
+            }
+        return response()->json($resp);
     }
 
 }
